move lib/logging.js => lib/logging/logging.js
