+++ bin/browserid
@@ -22,7 +22,7 @@ httputils = require('../lib/httputils.js'),
 db = require('../lib/db.js'),
 config = require('../lib/configuration.js'),
 heartbeat = require('../lib/heartbeat.js'),
-logger = require('../lib/logging.js').logger,
+logger = require('../lib/logging/logging.js').logger,
 shutdown = require('../lib/shutdown'),
 primary = require('../lib/primary'),
 addP3PHeader = require('../lib/p3p.js'),
