+++ bin/dbwriter
@@ -17,7 +17,7 @@ wsapi = require('../lib/wsapi.js'),
 db = require('../lib/db.js'),
 config = require('../lib/configuration.js'),
 heartbeat = require('../lib/heartbeat.js'),
-logger = require('../lib/logging.js').logger,
+logger = require('../lib/logging/logging.js').logger,
 shutdown = require('../lib/shutdown'),
 cachify = require('connect-cachify'),
 assets = require('../lib/static_resources').all,
