+++ bin/static
@@ -27,7 +27,7 @@ httputils = require('../lib/httputils.js'),
 db = require('../lib/db.js'),
 config = require('../lib/configuration.js'),
 heartbeat = require('../lib/heartbeat.js'),
-logger = require('../lib/logging.js').logger,
+logger = require('../lib/logging/logging.js').logger,
 views = require('../lib/static/views.js'),
 toobusy = require('../lib/busy_middleware.js'),
 addP3PHeader = require('../lib/p3p.js'),
