+++ bin/verifier
@@ -16,7 +16,7 @@ express = require('express'),
 computecluster = require('compute-cluster'),
 metrics = require('../lib/logging/metrics'),
 heartbeat = require('../lib/heartbeat'),
-logger = require('../lib/logging').logger,
+logger = require('../lib/logging/logging').logger,
 config = require('../lib/configuration'),
 shutdown = require('../lib/shutdown'),
 booleanQuery = require('../lib/boolean-query'),
