+++ lib/bcrypt.js
@@ -5,7 +5,7 @@
 const
 path = require('path'),
 computecluster = require('compute-cluster'),
-logger = require('../lib/logging.js').logger,
+logger = require('../lib/logging/logging.js').logger,
 bcrypt = require('bcrypt'),
 config = require('./configuration.js');
