+++ lib/browserid/fake_verification.js
@@ -12,7 +12,7 @@ const
 configuration = require('../configuration.js'),
 url = require('url'),
 db = require('../db.js'),
-logger = require('../logging.js').logger,
+logger = require('../logging/logging.js').logger,
 wsapi = require('../wsapi');
 
 logger.warn("HEAR YE: Fake verfication enabled, aceess via /wsapi/fake_verification?email=foo@bar.com");
