+++ lib/configuration.js
@@ -412,6 +412,6 @@ module.exports.performSubstitution = function(app) {
 
 // log the process_type
 process.nextTick(function() {
-  var logging = require("./logging.js").logger;
+  var logging = require("./logging/logging.js").logger;
   logging.info("process type is " + conf.get("process_type"));
 });
