+++ lib/db.js
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-var logger = require('./logging.js').logger;
+var logger = require('./logging/logging.js').logger;
 const config = require('./configuration.js');
 
 var driver;
