+++ lib/db/dbutils.js
@@ -3,7 +3,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 const
-logger = require('../logging.js').logger,
+logger = require('../logging/logging.js').logger,
 primary = require('../primary.js');
 
 // Callback has the signature `function (type) {}`
