+++ lib/db/json.js
@@ -13,7 +13,7 @@ path = require('path'),
 fs = require('fs'),
 secrets = require('../secrets.js'),
 jsel = require('JSONSelect'),
-logger = require('../logging.js').logger,
+logger = require('../logging/logging.js').logger,
 configuration = require('../configuration.js'),
 primary = require('../primary.js'),
 temp = require('temp');
@@ -209,7 +209,7 @@ exports.userOwnsEmail = function(uid, email, cb) {
     cb(null, owned);
   });
 };
- 
+
 function addEmailToAccount(userID, email, type, cb) {
   // validate 'type' isn't bogus
   if ([ 'secondary', 'primary' ].indexOf(type) === -1) {
