+++ lib/db/mysql.js
@@ -34,7 +34,7 @@ const
 conf = require('../configuration.js'),
 dbutils = require('./dbutils.js'),
 mysql = require('./mysql_wrapper.js'),
-logger = require('../logging.js').logger,
+logger = require('../logging/logging.js').logger,
 primary = require('../primary.js'),
 secrets = require('../secrets.js');
