+++ lib/email.js
@@ -8,7 +8,7 @@ fs = require('fs'),
 path = require('path'),
 ejs = require('ejs'),
 config = require('./configuration.js'),
-logger = require('./logging.js').logger,
+logger = require('./logging/logging.js').logger,
 url = require('url'),
 cachify = require('connect-cachify').cachify;
