+++ lib/generate_code_version.js
@@ -7,7 +7,7 @@ const
 fs = require('fs'),
 path = require('path'),
 mkdirp = require('mkdirp'),
-logger = require('../lib/logging').logger,
+logger = require('../lib/logging/logging').logger,
 version = require('../lib/version.js');
 
 const BUILD_DIR_PATH = path.join(__dirname, "..", "resources", "static", "build");
