+++ lib/heartbeat.js
@@ -4,7 +4,7 @@
 
 const
 urlparse = require('urlparse'),
-logger = require('./logging.js').logger,
+logger = require('./logging/logging.js').logger,
 url = require('url');
 
 // the path that heartbeats live at
