+++ lib/http_forward.js
@@ -6,7 +6,7 @@ const
 url = require('url'),
 http = require('http'),
 https = require('https'),
-logger = require('./logging.js').logger,
+logger = require('./logging/logging.js').logger,
 querystring = require('querystring');
 
 var global_forward_timeout;
