+++ lib/i18n_client_check.js
@@ -18,7 +18,7 @@
 var config = require('./configuration.js'),
     fs = require('fs'),
     i18n = require('i18n-abide'),
-    logger = require('./logging.js').logger,
+    logger = require('./logging/logging.js').logger,
     path = require('path'),
     util = require('util');
