+++ lib/keysigner/ca.js
@@ -7,7 +7,7 @@
 var jwcrypto = require('jwcrypto'),
     cert = jwcrypto.cert,
     secrets = require('../secrets.js'),
-    logger = require('../logging.js').logger;
+    logger = require('../logging/logging.js').logger;
 
 // load up the right algorithms
 require("jwcrypto/lib/algs/rs");
@@ -33,7 +33,7 @@ function certify(hostname, email, publicKey, expiration, unverified, cb) {
 
   var principal = {};
   principal[ unverified ? 'unverified-email' : 'email' ] = email;
-  
+
 
   cert.sign({publicKey: publicKey, principal: principal},
             {issuer: hostname, issuedAt: new Date(), expiresAt: expiration},
