+++ lib/logging/logging.js
@@ -16,9 +16,9 @@
 
 const
 winston = require("winston"),
-FileLogger = require('./logging/winston-file'),
-HekaLogger = require('./logging/winston-heka'),
-StatsdLogger = require('./logging/winston-statsd');
+FileLogger = require('./winston-file'),
+HekaLogger = require('./winston-heka'),
+StatsdLogger = require('./winston-statsd');
 
 
 exports.logger = new (winston.Logger)({
