+++ lib/logging/statsd.js
@@ -5,7 +5,7 @@
 const
 baseExceptions = require('../baseExceptions'),
 config = require('../configuration'),
-logger = require('../logging').logger,
+logger = require('./logging').logger,
 statsd_request_logger = require("connect-logger-statsd");
 
 var StatsD = false;
