+++ lib/primary.js
@@ -9,7 +9,7 @@
 const
 https = require('https'),
 http = require('http'),
-logger = require('./logging.js').logger,
+logger = require('./logging/logging.js').logger,
 urlparse = require('urlparse'),
 jwcrypto = require("jwcrypto"),
 config = require("./configuration.js"),
