+++ lib/shutdown.js
@@ -7,7 +7,7 @@
  * signals.
  */
 
-const logger = require("./logging.js").logger;
+const logger = require("./logging/logging.js").logger;
 
 const MAX_WAIT_MS = 10000;
 const MAX_NICE_END_MS = 5000;
