+++ lib/static/views.js
@@ -5,7 +5,7 @@
 const
 metrics = require('../logging/metrics.js'),
 url = require('url'),
-logger = require('../logging.js').logger,
+logger = require('../logging/logging.js').logger,
 fs = require('fs'),
 connect = require('connect'),
 config = require('../configuration.js'),
