+++ lib/validate.js
@@ -13,7 +13,7 @@
 //   });
 
 const
-logger = require('./logging.js').logger,
+logger = require('./logging/logging.js').logger,
 httputils = require('./httputils.js'),
 check = require('validator').check,
 url = require('url');
