+++ lib/verifier/certassertion.js
@@ -8,7 +8,7 @@ https = require("https"),
 url = require("url"),
 jwcrypto = require("jwcrypto"),
 config = require("../configuration.js"),
-logger = require("../logging.js").logger,
+logger = require("../logging/logging.js").logger,
 secrets = require('../secrets.js'),
 primary = require('../primary.js'),
 urlparse = require('urlparse');
@@ -159,7 +159,7 @@ function verify(assertion, audience, forceIssuer, allowUnverified, successCB, er
         email = principal[UNVERIFIED_EMAIL];
         verified = false;
       }
-      
+
       if (!email) {
         return errorCB("missing email");
       }
