+++ lib/version.js
@@ -10,7 +10,7 @@
 const
 fs = require('fs'),
 path = require('path'),
-logger = require('./logging.js').logger,
+logger = require('./logging/logging.js').logger,
 spawn = require('child_process').spawn,
 secrets = require('./secrets.js'),
 config = require('./configuration');
