+++ lib/wsapi.js
@@ -19,7 +19,7 @@ sessions = require('client-sessions'),
 express = require('express'),
 secrets = require('./secrets'),
 config = require('./configuration'),
-logger = require('./logging.js').logger,
+logger = require('./logging/logging.js').logger,
 httputils = require('./httputils.js'),
 forward = require('./http_forward.js').forward,
 url = require('url'),
