+++ lib/wsapi/account_cancel.js
@@ -5,7 +5,7 @@
 const
 db = require('../db.js'),
 wsapi = require('../wsapi'),
-logger = require('../logging.js').logger,
+logger = require('../logging/logging.js').logger,
 cef_logger = require('../logging/cef_logger').getInstance();
 
 exports.method = 'post';
