+++ lib/wsapi/add_email_with_assertion.js
@@ -6,7 +6,7 @@ const
 db = require('../db.js'),
 wsapi = require('../wsapi.js'),
 httputils = require('../httputils'),
-logger = require('../logging.js').logger,
+logger = require('../logging/logging.js').logger,
 querystring = require('querystring'),
 primary = require('../primary.js'),
 http = require('http'),
