+++ lib/wsapi/address_info.js
@@ -9,7 +9,7 @@ wsapi = require('../wsapi.js'),
 httputils = require('../httputils.js'),
 url = require('url'),
 urlparse = require('urlparse'),
-logger = require('../logging.js').logger,
+logger = require('../logging/logging.js').logger,
 config = require('../configuration.js');
 
 // type: {primary, secondary} // indicates email type right now
