+++ lib/wsapi/authenticate_user.js
@@ -6,7 +6,7 @@ const
 db = require('../db.js'),
 wsapi = require('../wsapi.js'),
 httputils = require('../httputils'),
-logger = require('../logging.js').logger,
+logger = require('../logging/logging.js').logger,
 bcrypt = require('../bcrypt'),
 http = require('http'),
 https = require('https'),
