+++ lib/wsapi/cert_key.js
@@ -5,7 +5,7 @@
 const
 db = require('../db.js'),
 httputils = require('../httputils'),
-logger = require('../logging.js').logger,
+logger = require('../logging/logging.js').logger,
 forward = require('../http_forward.js').forward,
 config = require('../configuration.js'),
 urlparse = require('urlparse'),
