+++ lib/wsapi/complete_email_confirmation.js
@@ -10,7 +10,7 @@
 
 const
 db = require('../db.js'),
-logger = require('../logging.js').logger,
+logger = require('../logging/logging.js').logger,
 wsapi = require('../wsapi.js'),
 bcrypt = require('../bcrypt.js'),
 httputils = require('../httputils.js');
