+++ lib/wsapi/complete_transition.js
@@ -4,7 +4,7 @@
 
 const
 db = require('../db.js'),
-logger = require('../logging.js').logger,
+logger = require('../logging/logging.js').logger,
 wsapi = require('../wsapi.js'),
 bcrypt = require('../bcrypt.js'),
 httputils = require('../httputils.js'),
