+++ lib/wsapi/complete_user_creation.js
@@ -6,7 +6,7 @@ const
 db = require('../db.js'),
 wsapi = require('../wsapi.js'),
 httputils = require('../httputils'),
-logger = require('../logging.js').logger,
+logger = require('../logging/logging.js').logger,
 bcrypt = require('../bcrypt'),
 config = require('../configuration'),
 wsapiutils = require('../wsapiutils');
