+++ lib/wsapi/create_account_with_assertion.js
@@ -7,7 +7,7 @@ db = require('../db.js'),
 wsapi = require('../wsapi.js'),
 httputils = require('../httputils'),
 primary = require('../primary.js'),
-logger = require('../logging.js').logger;
+logger = require('../logging/logging.js').logger;
 
 exports.method = 'post';
 exports.writes_db = true;
