+++ lib/wsapi/email_for_token.js
@@ -5,7 +5,7 @@
 const
 db = require('../db.js'),
 httputils = require('../httputils.js'),
-logger = require('../logging.js').logger;
+logger = require('../logging/logging.js').logger;
 
 /* Given a verification secret (a "token", delivered via email), return the
  * email address associated with this token.
