+++ lib/wsapi/interaction_data.js
@@ -6,7 +6,7 @@ const coarse = require('../coarse_user_agent_parser'),
       config = require('../configuration.js'),
       http = require('http'),
       https = require('https'),
-      logger = require('../logging.js').logger,
+      logger = require('../logging/logging.js').logger,
       querystring = require('querystring'),
       und = require('underscore'),
       urlparse = require('urlparse'),
@@ -47,7 +47,7 @@ var store = function (kpi_json, cb) {
     db_url = urlparse(config.get('kpi_backend_db_url'));
 
     http_proxy = config.has('http_proxy') ? config.get('http_proxy') : null;
-    
+
     if (http_proxy && http_proxy.port && http_proxy.host) {
       options = {
         host: http_proxy.host,
@@ -73,23 +73,23 @@ var store = function (kpi_json, cb) {
           'Content-Length': post_data.length
         }
       };
-      
+
       if (db_url.port) {
         options.port = db_url.port;
       }
-      
+
       var protocol = (db_url.scheme === 'https') ? https : http;
       kpi_req = protocol.request(options);
     }
-    
+
     kpi_req.on('response', function(res) {
       if (res.statusCode !== 201) {
         logger.warn('KPI Backend (or proxy) response code is not 201: ' + res.statusCode);
       } else {
         logger.info('interaction data successfully posted to KPI Backend');
-      }  
+      }
     });
-    
+
     kpi_req.on('error', function (e) {
       // TODO statsd counter
       logger.error('KPI Backend request error: ' + e.message);
