+++ lib/wsapi/list_emails.js
@@ -4,7 +4,7 @@
 
 const
 db = require('../db.js'),
-logger = require('../logging.js').logger,
+logger = require('../logging/logging.js').logger,
 wsapi = require('../wsapi.js');
 
 // returns a list of emails owned by the user:
