+++ lib/wsapi/remove_email.js
@@ -6,7 +6,7 @@ const
 db = require('../db.js'),
 wsapi = require('../wsapi'),
 httputils = require('../httputils'),
-logger = require('../logging.js').logger,
+logger = require('../logging/logging.js').logger,
 cef_logger = require('../logging/cef_logger').getInstance();
 
 exports.method = 'post';
