+++ lib/wsapi/saw_idp.js
@@ -12,7 +12,7 @@ db = require('../db.js'),
 wsapi = require('../wsapi.js'),
 httputils = require('../httputils'),
 primary = require('../primary.js'),
-logger = require('../logging.js').logger;
+logger = require('../logging/logging.js').logger;
 
 exports.method = 'get';
 exports.writes_db = true;
