+++ lib/wsapi/session_context.js
@@ -4,7 +4,7 @@
 
 const
 db = require('../db.js'),
-logger = require('../logging.js').logger,
+logger = require('../logging/logging.js').logger,
 crypto = require('crypto'),
 wsapi = require('../wsapi.js'),
 secrets = require('../secrets.js'),
