+++ lib/wsapi/stage_reverify.js
@@ -6,7 +6,7 @@ const
 db = require('../db.js'),
 wsapi = require('../wsapi.js'),
 httputils = require('../httputils'),
-logger = require('../logging.js').logger,
+logger = require('../logging/logging.js').logger,
 email = require('../email.js'),
 config = require('../configuration');
