+++ lib/wsapi/transition_status.js
@@ -5,7 +5,7 @@
 const
 db = require('../db.js'),
 wsapi = require('../wsapi.js'),
-logger = require('../logging.js').logger,
+logger = require('../logging/logging.js').logger,
 httputils = require('../httputils.js');
 
 exports.method = 'get';
