+++ lib/wsapi/update_password.js
@@ -6,7 +6,7 @@ const
 db = require('../db.js'),
 wsapi = require('../wsapi.js'),
 httputils = require('../httputils'),
-logger = require('../logging.js').logger,
+logger = require('../logging/logging.js').logger,
 bcrypt = require('../bcrypt');
 
 exports.method = 'post';
