+++ lib/wsapi/used_address_as_primary.js
@@ -5,7 +5,7 @@
 const
 db = require('../db.js'),
 httputils = require('../httputils'),
-logger = require('../logging.js').logger,
+logger = require('../logging/logging.js').logger,
 primary = require('../primary.js'),
 wsapi = require('../wsapi.js');
 
@@ -32,7 +32,7 @@ exports.process = function (req, res) {
     // not same account? big fat error
     if (!owned) return httputils.badRequest(res, "that email does not belong to you");
     var domain = primary.domainFromEmail(email);
-        
+
     primary.checkSupport(domain, function (err, r) {
       var notPrimary = false;
       if (err) {
