+++ lib/wsapi/user_used_email_as.js
@@ -11,7 +11,7 @@ db = require('../db.js'),
 wsapi = require('../wsapi.js'),
 httputils = require('../httputils'),
 primary = require('../primary.js'),
-logger = require('../logging.js').logger;
+logger = require('../logging/logging.js').logger;
 
 exports.method = 'post';
 exports.writes_db = true;
